feat(sampling):  Add recommended steps tests and tests clean up (#36354)
