+++ tests/js/spec/views/settings/project/server-side-sampling/index.spec.tsx
@@ -1,48 +1,16 @@
-import {initializeOrg} from 'sentry-test/initializeOrg';
-import {render, screen} from 'sentry-test/reactTestingLibrary';
+import {render, screen, userEvent, within} from 'sentry-test/reactTestingLibrary';
 
-import {Project} from 'sentry/types';
-import {OrganizationContext} from 'sentry/views/organizationContext';
-import {RouteContext} from 'sentry/views/routeContext';
-import ServerSideSampling from 'sentry/views/settings/project/server-side-sampling';
+import * as modal from 'sentry/actionCreators/modal';
 import {SERVER_SIDE_SAMPLING_DOC_LINK} from 'sentry/views/settings/project/server-side-sampling/utils';
 
-export function getMockData({
-  projects,
-  access,
-}: {access?: string[]; projects?: Project[]} = {}) {
-  return initializeOrg({
-    ...initializeOrg(),
-    organization: {
-      ...initializeOrg().organization,
-      features: ['server-side-sampling'],
-      access: access ?? initializeOrg().organization.access,
-      projects,
-    },
-    projects,
-  });
-}
+import {getMockData, mockedProjects, TestComponent, uniformRule} from './utils';
 
 describe('Server-side Sampling', function () {
   it('renders onboarding promo', function () {
     const {router, organization, project} = getMockData();
 
     const {container} = render(
-      <RouteContext.Provider
-        value={{
-          router,
-          location: router.location,
-          params: {
-            orgId: organization.slug,
-            projectId: project.slug,
-          },
-          routes: [],
-        }}
-      >
-        <OrganizationContext.Provider value={organization}>
-          <ServerSideSampling project={project} />
-        </OrganizationContext.Provider>
-      </RouteContext.Provider>
+      <TestComponent router={router} organization={organization} project={project} />
     );
 
     expect(
@@ -103,21 +71,7 @@ describe('Server-side Sampling', function () {
     });
 
     const {container} = render(
-      <RouteContext.Provider
-        value={{
-          router,
-          location: router.location,
-          params: {
-            orgId: organization.slug,
-            projectId: project.slug,
-          },
-          routes: [],
-        }}
-      >
-        <OrganizationContext.Provider value={organization}>
-          <ServerSideSampling project={project} />
-        </OrganizationContext.Provider>
-      </RouteContext.Provider>
+      <TestComponent router={router} organization={organization} project={project} />
     );
 
     // Rule Panel Header
@@ -186,25 +140,159 @@ describe('Server-side Sampling', function () {
     });
 
     render(
-      <RouteContext.Provider
-        value={{
-          router,
-          location: router.location,
-          params: {
-            orgId: organization.slug,
-            projectId: project.slug,
-          },
-          routes: [],
-        }}
-      >
-        <OrganizationContext.Provider value={organization}>
-          <ServerSideSampling project={project} />
-        </OrganizationContext.Provider>
-      </RouteContext.Provider>
+      <TestComponent router={router} organization={organization} project={project} />
     );
 
     const deleteButtons = screen.getAllByLabelText('Delete');
     expect(deleteButtons[0]).not.toHaveAttribute('disabled'); // eslint-disable-line jest-dom/prefer-enabled-disabled
     expect(deleteButtons[1]).toHaveAttribute('disabled'); // eslint-disable-line jest-dom/prefer-enabled-disabled
+  });
+
+  it('display "update sdk versions" alert and open "recommended next step" modal', async function () {
+    jest.spyOn(modal, 'openModal');
+
+    const {organization, projects, router} = getMockData({
+      projects: mockedProjects,
+    });
+
+    render(
+      <TestComponent
+        organization={organization}
+        project={projects[2]}
+        router={router}
+        withModal
+      />
+    );
+
+    const recommendedSdkUpgradesAlert = await screen.findByTestId(
+      'recommended-sdk-upgrades-alert'
+    );
+
+    expect(
+      within(recommendedSdkUpgradesAlert).getByText(
+        'To keep a consistent amount of transactions across your applications multiple services, we recommend you update the SDK versions for the following projects:'
+      )
+    ).toBeInTheDocument();
+
+    expect(
+      within(recommendedSdkUpgradesAlert).getByRole('link', {
+        name: mockedProjects[1].slug,
+      })
+    ).toHaveAttribute(
+      'href',
+      `/organizations/org-slug/projects/sentry/?project=${mockedProjects[1].id}`
+    );
+
+    // Open Modal
+    userEvent.click(
+      within(recommendedSdkUpgradesAlert).getByRole('button', {
+        name: 'Learn More',
+      })
+    );
+
+    expect(
+      await screen.findByRole('heading', {name: 'Recommended next steps\u2026'})
+    ).toBeInTheDocument();
+  });
+
+  it('Open activate modal', async function () {
+    const {router, project, organization} = getMockData({
+      projects: [
+        TestStubs.Project({
+          dynamicSampling: {
+            rules: [
+              {
+                sampleRate: 1,
+                type: 'trace',
+                active: false,
+                condition: {
+                  op: 'and',
+                  inner: [],
+                },
+                id: 1,
+              },
+            ],
+          },
+        }),
+      ],
+    });
+
+    render(
+      <TestComponent
+        organization={organization}
+        project={project}
+        router={router}
+        withModal
+      />
+    );
+
+    // Open Modal
+    userEvent.click(screen.getByLabelText('Activate Rule'));
+
+    expect(
+      await screen.findByRole('heading', {name: 'Activate Rule'})
+    ).toBeInTheDocument();
+  });
+
+  it('Open specific conditions modal', async function () {
+    jest.spyOn(modal, 'openModal');
+
+    const {router, project, organization} = getMockData({
+      projects: [
+        TestStubs.Project({
+          dynamicSampling: {
+            rules: [
+              {
+                sampleRate: 1,
+                type: 'trace',
+                active: false,
+                condition: {
+                  op: 'and',
+                  inner: [],
+                },
+                id: 1,
+              },
+            ],
+          },
+        }),
+      ],
+    });
+
+    render(
+      <TestComponent
+        organization={organization}
+        project={project}
+        router={router}
+        withModal
+      />
+    );
+
+    // Open Modal
+    userEvent.click(screen.getByLabelText('Add Rule'));
+
+    expect(await screen.findByRole('heading', {name: 'Add Rule'})).toBeInTheDocument();
+  });
+
+  it('does not let user add without permissions', async function () {
+    const {organization, router, project} = getMockData({
+      projects: [
+        TestStubs.Project({
+          dynamicSampling: {
+            rules: [uniformRule],
+          },
+        }),
+      ],
+      access: [],
+    });
+
+    render(
+      <TestComponent organization={organization} project={project} router={router} />
+    );
+
+    expect(screen.getByRole('button', {name: 'Add Rule'})).toBeDisabled();
+    userEvent.hover(screen.getByText('Add Rule'));
+    expect(
+      await screen.findByText("You don't have permission to add a rule")
+    ).toBeInTheDocument();
   });
 });
